requirements.txt
streamlit>=1.30.0
google-genai>=1.0.0
fpdf2>=2.8.0
beautifulsoup4>=4.12.0
requests>=2.30.0
python-dotenv>=1.0.0
